UIUX_Auditor/requirements.txt
Flask
openai
python-dotenv
requests
beautifulsoup4
playwright
